# server/controllers/Uberlist_db/generated/SeekerControllerGenerated.js
/**
 * 
 * 
  _____                      _              _ _ _     _   _     _        __ _ _      
 |  __ \                    | |            | (_) |   | | | |   (_)      / _(_) |     
 | |  | | ___    _ __   ___ | |_    ___  __| |_| |_  | |_| |__  _ ___  | |_ _| | ___ 
 | |  | |/ _ \  | '_ \ / _ \| __|  / _ \/ _` | | __| | __| '_ \| / __| |  _| | |/ _ \
 | |__| | (_) | | | | | (_) | |_  |  __/ (_| | | |_  | |_| | | | \__ \ | | | | |  __/
 |_____/ \___/  |_| |_|\___/ \__|  \___|\__,_|_|\__|  \__|_| |_|_|___/ |_| |_|_|\___|
 
 * DO NOT EDIT THIS FILE!! 
 * 
 *  TO CUSTOMIZE SeekerControllerGenerated.js PLEASE EDIT ../SeekerController.js
 * 
 *  -- THIS FILE WILL BE OVERWRITTEN ON THE NEXT SKAFFOLDER'S CODE GENERATION --
 * 
 */
import Properties from "../../../properties";
import SeekerModel from "../../../models/Uberlist_db/SeekerModel";
import ErrorManager from "../../../classes/ErrorManager";
import { authorize } from "../../../security/SecurityManager";
import SeekerController from "../SeekerController";

const generatedControllers = {
  /**
   * Init routes
   */
  init: router => {
    const baseUrl = `${Properties.api}/seeker`;
    router.post(baseUrl + "", authorize([]), SeekerController.create);
    router.delete(baseUrl + "/:id", authorize([]), SeekerController.delete);
    router.get(baseUrl + "/:id", authorize([]), SeekerController.get);
    router.get(baseUrl + "", authorize([]), SeekerController.list);
    router.post(baseUrl + "/:id", authorize([]), SeekerController.update);
  },


  // CRUD METHODS


  /**
  * SeekerModel.create
  *   @description CRUD ACTION create
  *   @param Seeker obj Object to insert
  *
  */
  create: async (req, res) => {
    try {
      const result = await SeekerModel.create(req.body);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * SeekerModel.delete
  *   @description CRUD ACTION delete
  *   @param ObjectId id Id
  *
  */
  delete: async (req, res) => {
    try {
      const result = await SeekerModel.delete(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * SeekerModel.get
  *   @description CRUD ACTION get
  *   @param ObjectId id Id Seeker
  *   @returns Seeker
  *
  */
  get: async (req, res) => {
    try {
      const result = await SeekerModel.get(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * SeekerModel.list
  *   @description CRUD ACTION list
  *
  */
  list: async (req, res) => {
    try {
      const result = await SeekerModel.list();
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  
  /**
  * SeekerModel.update
  *   @description CRUD ACTION update
  *   @param ObjectId id Id Seeker
  *   @returns Seeker
  *
  */
  update: async (req, res) => {
    try {
      const result = await SeekerModel.update(req.body);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  
  // Custom APIs

};

export default generatedControllers;